Separated each service by the roles that the methods within it play
