+++ src/Invoice_Gen.Services/ClientServices/ICreateClients.cs
@@ -1,0 +1,6 @@
+﻿namespace InvoiceGen.Services.ClientServices;
+
+public interface ICreateClients
+{
+    Task<int> CreateNewClient(ClientCreationModel inputClient);
+}
