+++ src/Invoice_Gen.Services/ClientServices/IDeleteClients.cs
@@ -1,0 +1,6 @@
+﻿namespace InvoiceGen.Services.ClientServices;
+
+public interface IDeleteClients
+{
+    Task DeleteClient(int clientId);
+}
