+++ src/Invoice_Gen.Services/ClientServices/IGetClients.cs
@@ -1,0 +1,7 @@
+﻿namespace InvoiceGen.Services.ClientServices;
+
+public interface IGetClients
+{
+    List<ClientViewModel> GetClients();
+    ClientViewModel? GetById(int id);
+}
