+++ src/Invoice_Gen.Services/ClientServices/IPageClients.cs
@@ -1,0 +1,6 @@
+﻿namespace InvoiceGen.Services.ClientServices;
+
+public interface IPageClients
+{
+    PagedResponse<ClientViewModel> GetPage(int pageNumber, int pageSize = 10);
+}
