--- src/Invoice_Gen.Services/ILineItemService.cs
@@ -1,8 +1,0 @@
-﻿using Invoice_Gen.ViewModels;
-
-namespace InvoiceGen.Services;
-
-public interface ILineItemService
-{
-    LineItemViewModel? GetById(int id);
-}
