+++ src/Invoice_Gen.Services/InvoiceServices/ICreateInvoices.cs
@@ -1,0 +1,7 @@
+﻿namespace InvoiceGen.Services.InvoiceServices;
+
+public interface ICreateInvoices
+{
+    Task<int> CreateNewInvoice(InvoiceCreateModel newInvoice);
+
+}
