+++ src/Invoice_Gen.Services/InvoiceServices/IDeleteInvoices.cs
@@ -1,0 +1,7 @@
+﻿namespace InvoiceGen.Services.InvoiceServices;
+
+public interface IDeleteInvoices
+{
+    Task DeleteInvoice(int invoiceId);
+
+}
