+++ src/Invoice_Gen.Services/InvoiceServices/IGetInvoices.cs
@@ -1,0 +1,7 @@
+﻿namespace InvoiceGen.Services.InvoiceServices;
+
+public interface IGetInvoices
+{
+    List<InvoiceViewModel> GetInvoices();
+    InvoiceViewModel? GetById(int id);
+}
