+++ src/Invoice_Gen.Services/InvoiceServices/IPageInvoices.cs
@@ -1,0 +1,6 @@
+﻿namespace InvoiceGen.Services.InvoiceServices;
+
+public interface IPageInvoices
+{
+    PagedResponse<InvoiceViewModel> GetPage(int pageNumber, int pageSize = 10); 
+}
