+++ src/Invoice_Gen.Services/LineItems/IGetLineItems.cs
@@ -1,0 +1,6 @@
+﻿namespace InvoiceGen.Services;
+
+public interface IGetLineItems
+{
+    LineItemViewModel? GetById(int id);
+}
