+++ src/Invoice_Gen.Services/Usings.cs
@@ -1,0 +1,5 @@
+﻿global using Invoice_Gen.Domain;
+global using Invoice_Gen.Domain.Models;
+global using Invoice_Gen.Mappers;
+global using Invoice_Gen.ViewModels;
+global using Microsoft.Extensions.Logging;
