+++ src/Invoice_Gen.WebApi/Controllers/InvoiceController.cs
@@ -1,5 +1,6 @@
 ﻿using System.Net.Mime;
 using InvoiceGen.Services;
+using InvoiceGen.Services.InvoiceServices;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Invoice_Gen.WebApi.Controllers;
@@ -9,13 +10,20 @@ namespace Invoice_Gen.WebApi.Controllers;
 [Produces(MediaTypeNames.Application.Json)]
 public class InvoiceController : ControllerBase
 {
-    private readonly IInvoiceService _invoiceService;
+    private readonly IGetInvoices _invoiceGetter;
+    private readonly ICreateInvoices _invoiceCreator;
+    private readonly IDeleteInvoices _invoiceDeleter;
+    private readonly IPageInvoices _pageInvoices;
     private readonly ILogger<InvoiceController> _logger;
 
-    public InvoiceController(ILogger<InvoiceController> logger, IInvoiceService invoiceService)
+    public InvoiceController(ILogger<InvoiceController> logger, IPageInvoices pageInvoices,
+        IGetInvoices invoiceGetter, ICreateInvoices invoiceCreator, IDeleteInvoices invoiceDeleter)
     {
         _logger = logger;
-        _invoiceService = invoiceService;
+        _pageInvoices = pageInvoices;
+        _invoiceGetter = invoiceGetter;
+        _invoiceCreator = invoiceCreator;
+        _invoiceDeleter = invoiceDeleter;
     }
 
     /// <summary>
@@ -30,7 +38,7 @@ public class InvoiceController : ControllerBase
     {
         using (_logger.BeginScope("Getting all invoices"))
         {
-            var invoices = _invoiceService.GetInvoices();
+            var invoices = _invoiceGetter.GetInvoices();
 
             _logger.LogInformation("Returning list of {InvoiceViewModel}", typeof(InvoiceViewModel));
             return new OkObjectResult(invoices);
@@ -51,7 +59,7 @@ public class InvoiceController : ControllerBase
     {
         using (_logger.BeginScope("Getting invoice data for {ID}", invoiceId))
         {
-            var invoice = _invoiceService.GetById(invoiceId);
+            var invoice = _invoiceGetter.GetById(invoiceId);
             if (invoice == null)
             {
                 _logger.LogInformation("Unable to find invoice record");
@@ -99,7 +107,7 @@ public class InvoiceController : ControllerBase
                     return new BadRequestResult();
             }
 
-            var pagedContent = _invoiceService.GetPage(pageNumber, pageSize);
+            var pagedContent = _pageInvoices.GetPage(pageNumber, pageSize);
             return new OkObjectResult(pagedContent);
         }
     }
@@ -121,7 +129,7 @@ public class InvoiceController : ControllerBase
         using (_logger.BeginScope("Request to create new client Invoice for client {ClientId} received",
                    newInvoice.ClientId))
         {
-            var response = await _invoiceService.CreateNewInvoice(newInvoice);
+            var response = await _invoiceCreator.CreateNewInvoice(newInvoice);
             if (response == default)
             {
                 return new StatusCodeResult(StatusCodes.Status500InternalServerError);
@@ -153,7 +161,7 @@ public class InvoiceController : ControllerBase
                 return new BadRequestResult();
             }
 
-            await _invoiceService.DeleteInvoice(invoiceId);
+            await _invoiceDeleter.DeleteInvoice(invoiceId);
 
             return new OkResult();
         }
