+++ tests/Invoice_Gen.WebApi.IntegrationTests/CustomWebApplicationFactory.cs
@@ -6,7 +6,6 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace Invoice_Gen.WebApi.IntegrationTests;
 
-[ExcludeFromCodeCoverage]
 public class CustomWebApplicationFactory : WebApplicationFactory<Program>
 {
     protected override void ConfigureWebHost(IWebHostBuilder builder)
