+++ tests/Invoice_Gen.WebApi.IntegrationTests/Invoice_Gen.WebApi.IntegrationTests.csproj
@@ -9,6 +9,10 @@
     </PropertyGroup>
 
     <ItemGroup>
+        <AssemblyAttribute Include="System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverageAttribute" />
+    </ItemGroup>
+
+    <ItemGroup>
         <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="7.0.8" />
         <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="7.0.8" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.3.2" />
