+++ tests/Invoice_Gen.WebApi.UnitTests/Helpers/ClientHelpers.cs
@@ -1,6 +1,5 @@
 ﻿namespace Invoice_Gen.WebApi.UnitTests.Helpers;
 
-[ExcludeFromCodeCoverage]
 public static class ClientHelpers
 {
     private static readonly Random Rng = new();
